20.category user 테이블 테스트

+++ src/test/java/com/example/study/repository/ItemRepositoryTest.java
@@ -18,7 +18,7 @@ public class ItemRepositoryTest {
     public void create() {
         Item item = new Item();
         item.setName("노트북");
-        item.setPrice(BigDecimal.valueOf(100000));
+        item.setPrice(100000);
         item.setContent("삼성 노트북");
 
         Item newItem = itemRepository.save(item);
